Extracted Notification to a separate file
