+++ BraveRewardsUI/Wallet/WalletViewController.swift
@@ -14,10 +14,13 @@ protocol WalletContentView: AnyObject {
 class WalletViewController: UIViewController, RewardsSummaryProtocol {
   
   let state: RewardsState
+  let notificationHandler: NotificationHandler
   
   init(state: RewardsState) {
     self.state = state
+    self.notificationHandler = NotificationHandler(ledger: state.ledger)
     super.init(nibName: nil, bundle: nil)
+    self.notificationHandler.delegate = self
   }
   
   @available(*, unavailable)
@@ -69,6 +72,7 @@ class WalletViewController: UIViewController, RewardsSummaryProtocol {
     
     reloadUIState()
     view.layoutIfNeeded()
+    notificationHandler.start()
   }
   
   override func viewWillAppear(_ animated: Bool) {
@@ -86,10 +90,6 @@ class WalletViewController: UIViewController, RewardsSummaryProtocol {
     if presentedViewController == nil {
       navigationController?.setNavigationBarHidden(false, animated: animated)
     }
-  }
-  
-  deinit {
-    stopNotificationService()
   }
   
   override func viewDidLayoutSubviews() {
@@ -130,172 +130,6 @@ class WalletViewController: UIViewController, RewardsSummaryProtocol {
     if preferredContentSize != newSize {
       preferredContentSize = newSize
     }
-  }
-  
-  // Notification
-  
-  private weak var currrentNotification: RewardsNotification?
-  
-  private func initiateNotificationService() {
-    // Add observer
-    NotificationCenter.default.addObserver(self, selector: #selector(notificationAdded(_:)), name: .init(rawValue: "BATBraveLedgerNotificationAdded"), object: nil)
-    loadNotification()
-  }
-  
-  private func stopNotificationService() {
-    NotificationCenter.default.removeObserver(self, name: .init(rawValue: "BATBraveLedgerNotificationAdded"), object: nil)
-  }
-  
-  @objc private func notificationAdded(_ notification: Notification) {
-    if currrentNotification == nil {
-      loadNotification()
-    }
-  }
-  
-  private func loadNotification() {
-    if let currrentNotification = currrentNotification {
-      state.ledger.clearNotification(currrentNotification)
-      self.currrentNotification = nil
-    }
-    
-    //.failedContribution, .invalid & .backupWallet are not supported right now
-    // TODO: Verify that notifications are ordered & they are displayed in FIFO.
-    let ignoredRewardNotificationTypes: [RewardsNotification.Kind] = [.failedContribution,
-                                                                      .invalid,
-                                                                      .backupWallet]
-    if let notification = state.ledger.notifications.first(where: {
-      !ignoredRewardNotificationTypes.contains($0.kind)
-    }) {
-      currrentNotification = notification
-      switch notification.kind {
-      case .autoContribute, .tipsProcessed, .verifiedPublisher:
-        addMessageNotification(notification: notification)
-      case .grant, .grantAds, .insufficientFunds, .pendingNotEnoughFunds, .adsLaunch:
-        addActionNotification(notification: notification)
-      default:
-        return
-      }
-    }
-  }
-  
-  private func addActionNotification(notification: RewardsNotification) {
-    let selector: Selector
-    let category: WalletActionNotification.Category
-    let body: String
-    
-    switch notification.kind {
-    case .grant:
-      body = Strings.NotificationGrantNotification
-      selector = #selector(tappedGrantsButton)
-      category = .grant
-    case .grantAds:
-      body = Strings.NotificationEarningsClaimDefault
-      selector = #selector(tappedGrantsButton)
-      category = .grant
-    case .insufficientFunds:
-      body = Strings.NotificationInsufficientFunds
-      selector = #selector(tappedAddFunds)
-      category = .insufficientFunds
-    case .pendingNotEnoughFunds:
-      body = Strings.NotificationPendingNotEnoughFunds
-      selector = #selector(tappedAddFunds)
-      category = .insufficientFunds
-    case .adsLaunch:
-      body = Strings.NotificationBraveAdsLaunchMsg
-      selector = #selector(tappedSettings)
-      category = .adsLaunch
-    default:
-      loadNotification()
-      return
-    }
-    let date = Date(timeIntervalSince1970: notification.dateAdded)
-    let notificationView = WalletActionNotificationView(
-      notification: WalletActionNotification(
-        category: category,
-        body: body,
-        date: date))
-    notificationView.closeButton.addTarget(self, action: #selector(tappedNotificationClose), for: .touchUpInside)
-    notificationView.actionButton.addTarget(self, action: selector, for: .touchUpInside)
-    self.walletView.setNotificationView(notificationView, animated: true)
-  }
-  
-  private func addMessageNotification(notification: RewardsNotification) {
-    var category: WalletNotification.Category = .contribute
-    let body: String
-    var alertType: WalletAlertNotification.Category?
-    switch notification.kind {
-    case .autoContribute:
-      if let result = notification.userInfo["result"] as? Int, let amount = notification.userInfo["amount"] as? String {
-        switch result {
-        case 0:
-          body = String.localizedStringWithFormat(Strings.NotificationContributeSuccess, amount)
-          category = .contribute
-        case 15:
-          body = Strings.NotificationAutoContributeNotEnoughFundsBody
-          alertType = .warning
-        case 16:
-          body = Strings.NotificationContributeTipError
-          alertType = .error
-        default:
-          body = Strings.NotificationContributeError
-          alertType = .error
-        }
-      } else {
-        loadNotification()
-        return
-      }
-    case .tipsProcessed:
-      body = Strings.NotificationTipsProcessedBody
-      category = .tipsProcessed
-    case .verifiedPublisher:
-      // FIXME: Verify key
-      if let name = notification.userInfo["name"] as? String {
-        body = String.localizedStringWithFormat(Strings.NotificationVerifiedPublisherBody, name)// publisher name"
-        category = .verifiedPublisher
-      } else {
-        loadNotification()
-        return
-      }
-    default:
-      loadNotification()
-      return
-    }
-    if let type = alertType {
-      showAlertNotification(type: type, message: body)
-      return
-    }
-    
-    let date = Date(timeIntervalSince1970: notification.dateAdded)
-    let notificationView = WalletNotificationView(
-      notification: WalletNotification(
-        category: category,
-        body: body,
-        date: date))
-    notificationView.closeButton.addTarget(self, action: #selector(tappedNotificationClose), for: .touchUpInside)
-    self.walletView.setNotificationView(notificationView, animated: true)
-  }
-  
-  private func showAlertNotification(type: WalletAlertNotification.Category, message: String) {
-    let notificationView = WalletAlertNotificationView(
-      notification: WalletAlertNotification(
-        category: type,
-        title: nil,
-        body: message
-      )
-    )
-    notificationView.closeButton.addTarget(self, action: #selector(tappedNotificationClose), for: .touchUpInside)
-    self.walletView.setNotificationView(notificationView, animated: true)
-  }
-  
-  @objc private func tappedNotificationClose() {
-    self.walletView.setNotificationView(nil, animated: true)
-    loadNotification()
-  }
-  
-  @objc private func tappedNotificationAction() {
-    // TODO: Do something with the notification
-    self.walletView.setNotificationView(nil, animated: true)
-    loadNotification()
   }
   
   // MARK: -
@@ -504,5 +338,29 @@ class WalletViewController: UIViewController, RewardsSummaryProtocol {
     default:
       assertionFailure()
     }
+  }
+}
+
+extension WalletViewController: NotificationHandlerDelegate {
+  func show(notificationView: WalletNotificationView) {
+    self.walletView.setNotificationView(notificationView, animated: true)
+  }
+  
+  func handleAction(notification: RewardsNotification) {
+    // TODO: Verify the action usage
+    switch notification.kind {
+    case .grant, .grantAds:
+      tappedGrantsButton()
+    case .pendingNotEnoughFunds, .insufficientFunds:
+      tappedAddFunds()
+    case .adsLaunch:
+      tappedSettings()
+    default:
+      return
+    }
+  }
+  
+  func hideNotifications() {
+    self.walletView.setNotificationView(nil, animated: true)
   }
 }
